Modal moved to its on folder
